+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+Unreleased
+----------
+* BREAKING CHANGE: ``market_watch`` will now use ``string`` as the dtype for ``ins_code`` index.
+
 v0.50.0
 -------
 * BREAKING CHANGE: ``Instrument.related_companies`` now adds an ``instrument.`` prefix to instrument related column names.